Rewrite sites_recodes.R and listPrimarySite()
For better visualization of primary sites, I rewrote the sites_recode.R
script that generates the data with the name of sites and recodes using
data.tree to print a tree structure of sites. Then, I rewrote the
listPrimarySites() function and primarySiteLookUp() to make better use
of the data tree.