Update buildSEERParser function and code (linting)
If the file_source parameter in buildSEERParser function is neither
"download" or "seerstat", the function would only print a warning but
would not stop. This was fixed.

Something else that I missed is that the buildSEERParser was not
returning the output.

Changes were also made to all files according to suggestions from
lintr::lint_package().