Fix no NA bug and vector bug
Due to the way [] subsetting works, removeFullNAfeatures would remove
all columns for a dataframe without a full NA column. That is, if I
executed the following line twice:
x <- x[, -which(colMeans(is.na(x)) == 1)]
All columns would be removed.

Besides, though the manual page mentioned you could pass a vector of
additional NA terms to the function, it was only working with one
additional term.