Create function findSingleValueFeatures()
Create function findSingleValueFeatures to help the user to find all
features that have an unique value for all rows in the dataset. This can
occur due to the way the user subsetted the dataset. If he is only
looking at cancer in breast site, the feature for cancer site will only
have 'Breast'. If he is only looking at tumors diagnosed in 2010, year
of diagnosis will only have one value which is 2010. There are several
reasons for an unique value in a feature but usually it's best to remove
such features from the dataset.